refactor(backend): split tmdb.ts into submodules + fetchList helper (#137)
routes/tmdb.ts was 443 lines with 21 handlers, 12 of which copy-pasted the
same pattern (parse lang/page → call service → flagNsfwFromDb → return).
Extracted a fetchList helper and split by concern:

- tmdb/helpers.ts — getLang, schemas, flagNsfwFromDb, fetchList
- tmdb/list.ts    — 12 paginated list endpoints (trending, popular, upcoming,
                    anime, now_playing, top_rated, airing_today, on_the_air,
                    recommendations movie/tv)
- tmdb/details.ts — single-resource lookups (movie/tv/person/collection)
- tmdb/discover.ts — search + discover by genre + flat discover
- tmdb/genres.ts  — genre-backdrops + per-mediaType genre list

tmdb.ts drops to 17 lines of orchestration. Adding a new list endpoint is
now a one-liner via fetchList.

NSFW flagging retains its best-effort catch but now logs to request.log.warn
on failure — a consistent prisma outage or schema drift no longer silently
degrades NSFW filtering to zero flags. The /discover/:mediaType flat route
drops its misleading querystring schema since the real param gate is the
ALLOWED_PARAMS whitelist inside the handler.

No behavior change on any endpoint. Unit tests for fetchList deferred to
when we add a proper Vitest harness.
